Update prompt
Ubuntu and Fedora nowadays both use a colorful prompt by default. Maybe
its time for us to do that as well. I am not exactly happy with the new
style of the prompt string but it is more in line with what Debian,
Fedora and Ubuntu ship with out of the box.